disasm: use a more relevant address type

+++ src/disasm.h
@@ -2,4 +2,4 @@
 
 #include <stdint.h>
 
-void disasm(const uint8_t *buffer, const int size, const int address, const int illegal);
+void disasm(const uint8_t *buffer, const int size, const uint16_t address, const int illegal);
